third_party/WebKit/LayoutTests/platform/mac-retina/external/wpt/service-workers/service-worker/ServiceWorkerGlobalScope/extendable-message-event.https-expected.txt

This is a testharness.js-based test.
FAIL Post an extendable message from a top-level client assert_equals: event `source` property `focused` expected true but got false
PASS Post an extendable message from a nested client
PASS Post loopback extendable messages
PASS Post extendable messages among service workers
Harness: the test ran to completion.
